knowgl/pyproject.toml
[tool.poetry]
name = "knowgl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.35.2"
ipykernel = "^6.27.0"
torchvision = "^0.16.1"
torchaudio = "^2.1.1"
nltk = "^3.8.1"
pyvis = "^0.3.2"
iprogress = "^0.4"
spacy = "^3.7.2"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl"}
neo4j = "^4.4.2"
fastcoref = "^2.1.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
language_detection = "pipeline.language_detection.main:main"
sentence_parser = "pipeline.sentence_parser.main:main"
translation = "pipeline.translation.main:main"
coreference_resolution = "pipeline.coreference_resolution.main:main"
pipeline = "pipeline.main:main"